# net, websocket: replace Crypto++ calls with GnuTLS