since we already depend on GnuTLS which provides the functions which we use Crypto++ for. so let's rewrite the Crypto++ function calls using the ones provided by GnuTLS for less dependencies.

Fixes #680
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>