@tchaikov 2023-04-04
i lost my virtualbox virtual machine which was setup for testing dpdk applications after messing up with my debian box,  not able to setup another at this moment. the testpmd application from dpdk works, but the httpd application from seastar fails even without my patch in this PR. probably something wrong with the NIC settings of the VM. will try harder later on.