@tchaikov 2023-12-26
v2:

- rebased against master HEAD.