@tchaikov 2023-12-26
i just setup my machine for testing this change with following step:

```console
# # follow the instructions at https://core.dpdk.org/doc/quick-start/
# echo 64 > /sys/devices/system/node/node0/hugepages/hugepages-2048kB/nr_hugepages
$ ./configure.py --enable-dpdk --cook dpdk --compiler clang++ --c-compiler clang --mode release
$ ninja -C build/release/ httpd
$ build/release/apps/httpd/httpd --dpdk-pmd
INFO  2023-12-26 14:05:50,006 seastar - Reactor backend: io_uring
INFO  2023-12-26 14:05:50,009 seastar - Perf-based stall detector creation failed (EACCESS), try setting /proc/sys/kernel/perf_event_paranoid to 1 or less to enable kernel backtraces: falling back to posix timer.
starting prometheus API server
Seastar HTTP server listening on port 10000 ...
^CStoppping HTTP server
```

in another console on the same machine:
```
$ git describe 
3.0.1-49-ga211dd5
$ ./wrk -c 10 -d 10 -t 10 http://localhost:10000
```

== before this change
```console
$ ./wrk -c 10 -d 10 -t 10 http://localhost:10000
Running 10s test @ http://localhost:10000
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    37.14us   82.80us   8.06ms   98.70%
    Req/Sec    29.75k     2.86k   37.07k    71.98%
  2989467 requests in 10.10s, 390.58MB read
Requests/sec: 295998.17
Transfer/sec:     38.67MB
```

== after this change
```console
$ ./wrk -c 10 -d 10 -t 10 http://localhost:10000                                                                                                                                                        14:11:09 [96/130]
Running 10s test @ http://localhost:10000
  10 threads and 10 connections                            
  Thread Stats   Avg      Stdev     Max   +/- Stdev                                                                    
    Latency    37.51us   84.38us   4.39ms   98.28%         
    Req/Sec    30.76k     3.26k   45.89k    73.98%
  3082189 requests in 10.10s, 402.70MB read                
Requests/sec: 305174.24                   
Transfer/sec:     39.87MB          
```

the change does not bring significant performance regression or break the functionality.