@tchaikov 2023-12-26
@scylladb/seastar-maint hello maintainers, i rebased the change and tested it locally. could you please take a look? sorry for the latency.